fixed regression for deleting items
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@1854 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/tasks/WorkbenchTask.java
@@ -1989,17 +1989,24 @@ public class WorkbenchTask extends BaseTask
                 Collection<WorkbenchTemplateMappingItem> newItems     = dlg.getNewItems();
                 //Collection<WorkbenchTemplateMappingItem> updatedItems = dlg.getUpdatedItems();
                 
-                WorkbenchTemplate workbenchTemplate = (WorkbenchTemplate)session.merge(wbTemplate);
                 session.beginTransaction();
                 
+                WorkbenchTemplate workbenchTemplate = (WorkbenchTemplate)session.merge(wbTemplate);
+                
                 Set<WorkbenchTemplateMappingItem> items = workbenchTemplate.getWorkbenchTemplateMappingItems();
-
-                for (WorkbenchTemplateMappingItem wbtmi : unusedItems)
+                for (WorkbenchTemplateMappingItem unusedItem : unusedItems)
                 {
-                    //log.debug("del ["+wbtmi.getCaption()+"]["+wbtmi.getWorkbenchTemplateMappingItemId().longValue()+"]");
-                    wbtmi.setWorkbenchTemplate(null);
-                    items.remove(wbtmi);
-                    session.delete(wbtmi) ;
+                    for (WorkbenchTemplateMappingItem wbtmi : items)
+                    {
+                        if (unusedItem.getWorkbenchTemplateMappingItemId().longValue() == wbtmi.getWorkbenchTemplateMappingItemId().longValue())
+                        {
+                            //log.debug("del ["+wbtmi.getCaption()+"]["+wbtmi.getWorkbenchTemplateMappingItemId().longValue()+"]");
+                            wbtmi.setWorkbenchTemplate(null);
+                            items.remove(wbtmi);
+                            session.delete(wbtmi);
+                            break;
+                        }
+                    }
                 }
                 for (WorkbenchTemplateMappingItem wbtmi : newItems)
                 {
